split UVCCameraTextureView into UVCCameraTextureView and AspectRatioTextureView
